# src/ads/ads.controller.ts
import {
	Controller,
	Get,
	Post,
	Body,
	Patch,
	Param,
	Delete,
	Query
} from '@nestjs/common'
import { CreateAdDto } from './dto/create-ad.dto'
import { Auth } from 'src/auth/decorators/auth.decorator'
import { CurrentUser } from 'src/auth/decorators/current-user.decorator'
import { AdsService } from './ads.service'
import { AdType } from 'prisma/generated/enums'
import { ApiBearerAuth, ApiOperation, ApiTags } from '@nestjs/swagger'

@ApiTags('Ads')
@Controller('ads')
export class AdsController {
	constructor(private readonly adsService: AdsService) {}

	@Post()
	@ApiBearerAuth('accessToken')
	@Auth()
	@ApiOperation({ summary: 'Создать объявление (требуется JWT)' })
	create(@CurrentUser('id') userId: string, @Body() dto: CreateAdDto) {
		return this.adsService.create(userId, dto)
	}

	@Get()
	@ApiOperation({ summary: 'Получить список объявлений' })
	findAll(@Query('type') type?: AdType) {
		return this.adsService.getAll(type)
	}

	@Get('by-slug/:slug')
	@ApiOperation({ summary: 'Получить объявление по слагу' })
	getBySlug(@Param('slug') slug: string) {
		return this.adsService.getBySlug(slug)
	}

	@Patch(':id')
	@Auth()
	@ApiOperation({ summary: 'Обновить объявление по id (только для автора)' })
	update(
		@CurrentUser('id') userId: string,
		@Param('id') id: string,
		@Body() dto: Partial<CreateAdDto>
	) {
		return this.adsService.update(userId, id, dto)
	}

	@Patch(':id/close')
	@Auth()
	@ApiOperation({ summary: 'Закрыть объявление по id' })
	close(@CurrentUser('id') userId: string, @Param('id') id: string) {
		return this.adsService.closeAd(userId, id)
	}

	@Patch(':id/reopen')
	@Auth()
	@ApiOperation({ summary: 'Вернуть объявление в статус Активно по id' })
	reopen(@CurrentUser('id') userId: string, @Param('id') id: string) {
		return this.adsService.reopenAd(userId, id)
	}

	@Delete(':id')
	@ApiBearerAuth('accessToken')
	@Auth()
	@ApiOperation({ summary: 'Удалить объявление (только ADMIN)' })
	remove(@CurrentUser('id') userId: string, @Param('id') id: string) {
		return this.adsService.remove(userId, id)
	}
}