# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/yandex":{"get":{"operationId":"AuthController_yandexAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/yandex/callback":{"get":{"operationId":"AuthController_yandexAuthCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/new-tokens":{"post":{"operationId":"AuthController_newTokens","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/verify-email":{"get":{"operationId":"AuthController_verifyEmail","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Auth"]}},"/api/auth/password-reset/request":{"post":{"operationId":"AuthController_requestReset","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Auth"]}},"/api/auth/password-reset/confirm":{"post":{"operationId":"AuthController_confirmReset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Auth"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Auth"]}},"/api/users/leaderboard":{"get":{"operationId":"UsersController_getLeaderboard","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/profile":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/full-profile":{"get":{"operationId":"UsersController_getFullProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/admin-stats":{"get":{"operationId":"UsersController_getAdminStats","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/get-activity":{"get":{"operationId":"UsersController_getActivity","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users":{"get":{"operationId":"UsersController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/profile/change-password":{"patch":{"operationId":"UsersController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/users/collect-waste":{"post":{"operationId":"UsersController_collectWaste","parameters":[],"responses":{"201":{"description":""}},"tags":["Users"]}},"/api/achievements/my":{"get":{"operationId":"AchievementsController_getMyAchievements","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Achievements"]}},"/api/achievements":{"get":{"operationId":"AchievementsController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Achievements"]},"post":{"operationId":"AchievementsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAchievementDto"}}}},"responses":{"201":{"description":""}},"tags":["Achievements"]}},"/api/achievements/{id}":{"put":{"operationId":"AchievementsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAchievementDto"}}}},"responses":{"200":{"description":""}},"tags":["Achievements"]},"delete":{"operationId":"AchievementsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Achievements"]}},"/api/map-points":{"get":{"operationId":"MapController_findAll","parameters":[{"name":"types","required":true,"in":"query","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Map"]},"post":{"operationId":"MapController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMapPointDto"}}}},"responses":{"201":{"description":""}},"tags":["Map"]}},"/api/map-points/pending":{"get":{"operationId":"MapController_getPendingPoints","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Map"]}},"/api/map-points/{id}":{"get":{"operationId":"MapController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Map"]},"put":{"operationId":"MapController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMapPointDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Map"]},"delete":{"operationId":"MapController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Map"]}},"/api/map-points/{id}/verify":{"patch":{"operationId":"MapController_verifyPoint","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Map"]}},"/api/waste":{"get":{"operationId":"WasteController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Waste"]},"post":{"operationId":"WasteController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWasteDto"}}}},"responses":{"201":{"description":"Успешно создано"}},"summary":"Создать объявление","tags":["Waste"]}},"/api/waste/{id}":{"delete":{"operationId":"WasteController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Waste"]}},"/api/media-upload":{"post":{"operationId":"MediaUploadController_uploadImage","parameters":[{"name":"folder","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["MediaUpload"]}},"/api/ads":{"post":{"operationId":"AdsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAdDto"}}}},"responses":{"201":{"description":""}},"security":[{"accessToken":[]}],"summary":"Создать объявление (требуется JWT)","tags":["Ads"]},"get":{"operationId":"AdsController_findAll","parameters":[{"name":"type","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"summary":"Получить список объявлений","tags":["Ads"]}},"/api/ads/by-slug/{slug}":{"get":{"operationId":"AdsController_getBySlug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"summary":"Получить объявление по слагу","tags":["Ads"]}},"/api/ads/{id}":{"patch":{"operationId":"AdsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Обновить объявление по id (только для автора)","tags":["Ads"]},"delete":{"operationId":"AdsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"accessToken":[]}],"summary":"Удалить объявление (только ADMIN)","tags":["Ads"]}},"/api/ads/{id}/close":{"patch":{"operationId":"AdsController_close","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Закрыть объявление по id","tags":["Ads"]}},"/api/ads/{id}/reopen":{"patch":{"operationId":"AdsController_reopen","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Вернуть объявление в статус Активно по id","tags":["Ads"]}},"/api/events":{"post":{"operationId":"EventsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"}}}},"responses":{"201":{"description":""}},"tags":["Events"]},"get":{"operationId":"EventsController_findAll","parameters":[{"name":"category","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}},{"name":"status","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Events"]}},"/api/events/by-slug/{slug}":{"get":{"operationId":"EventsController_getBySlug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Events"]}},"/api/events/{id}":{"put":{"operationId":"EventsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Events"]},"delete":{"operationId":"EventsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Events"]}},"/api/events/{id}/toggle-registration":{"post":{"operationId":"EventsController_toggleRegistration","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Events"]}},"/api/events/{id}/finish":{"post":{"operationId":"EventsController_finish","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Events"]}},"/api/messages/chats":{"get":{"operationId":"MessagesController_getChatList","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Messages"]}},"/api/messages":{"get":{"operationId":"MessagesController_getChatHistory","parameters":[{"name":"adId","required":true,"in":"query","schema":{"type":"string"}},{"name":"eventId","required":true,"in":"query","schema":{"type":"string"}},{"name":"participantId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Messages"]},"post":{"operationId":"MessagesController_sendMessage","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Messages"]}}},"info":{"title":"EcoNeighbor API","description":"Документация для фронтенда эко-платформы","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Введите accessToken для работы запроса","in":"header"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6},"name":{"type":"string"}},"required":["email","password","name"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6}},"required":["email","password"]},"ResetPasswordDto":{"type":"object","properties":{"token":{"type":"string"},"newPassword":{"type":"string"}},"required":["token","newPassword"]},"ChangePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string","minLength":6}},"required":["oldPassword","newPassword"]},"CreateAchievementDto":{"type":"object","properties":{"slug":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"iconUrl":{"type":"string"},"category":{"type":"string"},"unit":{"type":"object"},"requirementCount":{"type":"number"},"points":{"type":"number"}},"required":["slug","title","description","iconUrl","category","unit","requirementCount","points"]},"CreateMapPointDto":{"type":"object","properties":{"title":{"type":"string"},"address":{"type":"string"},"latitude":{"type":"number"},"longitude":{"type":"number"},"description":{"type":"string"},"isVerified":{"type":"boolean"},"type":{"type":"object"},"wasteIds":{"type":"array","items":{"type":"string"}}},"required":["title","address","latitude","longitude","wasteIds"]},"CreateWasteDto":{"type":"object","properties":{"name":{"type":"string","example":"Пластиковая бутылка","description":"Заголовок"}},"required":["name"]},"CreateAdDto":{"type":"object","properties":{"title":{"type":"string","example":"Отдам старый кактус","description":"Заголовок объявления"},"slug":{"type":"string","example":"otdam-stariy-kaktys-36sfs","description":"Slug объявления"},"description":{"type":"string","example":"Здоровый, колючий, самовывоз","description":"Описание"},"phone":{"type":"string","example":"+79999999999","description":"Телефон"},"type":{"type":"string","enum":["GIVEAWAY","RECYCLE","NEED_HELP"],"example":"GIVEAWAY"},"imageUrl":{"type":"string"}},"required":["title","slug","description","type"]},"CreateEventDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"category":{"type":"object"},"phone":{"type":"string"},"imageUrl":{"type":"string"},"date":{"type":"object"},"location":{"type":"string"},"latitude":{"type":"number"},"longitude":{"type":"number"},"maxParticipants":{"type":"number"}},"required":["title","description","category","date","location","latitude","longitude"]},"Object":{"type":"object","properties":{}}}}}